# Conflicts:
#	Plan/common/src/main/java/com/djrapitops/plan/PlanSystem.java
#	Plan/common/src/main/java/com/djrapitops/plan/delivery/rendering/pages/NetworkPage.java
#	Plan/common/src/main/java/com/djrapitops/plan/identification/Identifiers.java
#	Plan/common/src/main/java/com/djrapitops/plan/settings/locale/lang/JSLang.java
#	Plan/common/src/main/java/com/djrapitops/plan/storage/database/SQLDB.java
#	Plan/common/src/main/resources/assets/plan/locale/locale_CN.yml
#	Plan/common/src/main/resources/assets/plan/locale/locale_CS.yml
#	Plan/common/src/main/resources/assets/plan/locale/locale_EN.yml
#	Plan/common/src/main/resources/assets/plan/locale/locale_ES.yml
#	Plan/common/src/test/java/com/djrapitops/plan/delivery/export/ExportJSErrorRegressionTest.java
#	Plan/plugin/build.gradle
#	Plan/react/dashboard/public/index.html
#	Plan/react/dashboard/src/style/sb-admin-2.css
#	Plan/settings.gradle
#	Plan/sponge/src/main/java/com/djrapitops/plan/gathering/listeners/sponge/SpongeGMChangeListener.java

+++ Plan/common/src/main/java/com/djrapitops/plan/delivery/domain/mutators/RetentionData.java
@@ -46,8 +46,8 @@ public class RetentionData {
             totalPlayersOnline += retentionData.getOnlineOnJoin();
         }
 
-        double averageIndex = totalIndex / (double) size;
-        double averagePlayersOnline = totalPlayersOnline / (double) size;
+        double averageIndex = totalIndex / size;
+        double averagePlayersOnline = totalPlayersOnline / size;
 
         return new RetentionData(averageIndex, averagePlayersOnline);
     }
